+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 	  run as apache
 - add scripts: uninstall script, backup script and get-mizdb script
 - moved database scripts into `scripts` directory
+- changed update process:
+	- update now checks against GitHub API
+	- rebuilds containers after pulling the update
+	- applies migrations if necessary
 
 ### Fixed